Fix and optimize time left calculation and display
Prior to this, there was a "gap" in display of sprint time left, causing
the progress bar to display no data if remaining time was >0 and <7200.

The "time left" display strings have been modified to use %d and sprintf
to facilitate i18n.

This commit uses the new duration constants instead of hardcoded values
to determine what to display.